First Phase: WebHooks and Playlists

+++ modules/player/index.coffee
@@ -7,7 +7,7 @@ AudioHud = reload './hud'
 
 class PlayerModule extends BotModule
   init: =>
-    { @permissions, @getGuildData } = @engine
+    { @permissions, @getGuildData, @webHooks } = @engine
     @audioFilters = audioFilters
     @hud = new AudioHud @
     @moduleCommands = new AudioModuleCommands @
@@ -21,7 +21,17 @@ class PlayerModule extends BotModule
     }
     .asSeconds()
 
-  handleVideoInfo: (info, msg, args)=>
+  handleVideoInfo: (info, msg, args, silent=false)=>
+    # Check if playlist
+    if typeof info.forEach is 'function'
+      # Playlist
+      @webHooks.getForChannel(msg.channel, true)
+      .then (hook)=>
+        hook.execSlack @hud.addPlaylistWebhook(msg.author, info.length)
+      .catch (e)=>
+        msg.channel.sendMessage @hud.addPlaylist(msg.author, info.length)
+      # Iterate over all items
+      return info.forEach (v)=> @handleVideoInfo(v, msg, args, true)
     # Check if duration is valid
     duration = @parseTime info.duration
        
@@ -29,7 +39,8 @@ class PlayerModule extends BotModule
        (duration > 7200  and not @permissions.isAdmin(msg.author, msg.guild)) or
        (duration > 43200 and not @permissions.isOwner(msg.author))
        # Bot Owner = ∞
-      return msg.reply 'The requested song is too long.'
+      return msg.reply 'The requested song is too long.' if not silent
+      return
 
     # Get filters
     filters = []
@@ -41,10 +52,12 @@ class PlayerModule extends BotModule
           filter = audioFilters.getFilter f[0], f[1]
           if filter
             if filter.isAdminOnly() and not @permissions.isDJ msg.author, msg.guild
-              return msg.reply "#{filter} is only for DJs."
+              return msg.reply "#{filter} is only for DJs." if not silent
+              return
             valErr = filter.validate()
             if valErr
-              return msg.reply "#{filter} - #{valErr}"
+              return msg.reply "#{filter} - #{valErr}" if not silent
+              return
             filters.push filter
 
     omsg = undefined
@@ -64,6 +77,7 @@ class PlayerModule extends BotModule
       filters: filters
       path: info.url
       sauce: info.webpage_url
+      thumbnail: info.thumbnail
     }
     # Set events
     durationstr = if isFinite(qI.duration) then moment.utc(qI.duration * 1000).format("HH:mm:ss") else '∞'
@@ -83,16 +97,24 @@ class PlayerModule extends BotModule
             setTimeout (->m.delete()), 15000
           audioPlayer.clean true
       ), 100
-      
-    msg.channel.sendMessage @hud.addItem msg.guild, qI.requestedBy, qI, queue.items.length+1
-    .then (m)=>
-      omsg = m;
-      setTimeout ->
-        if omsg
-          omsg.delete()
-          omsg = null
-      , 15000
-      msg.delete()
+    
+    if not silent
+      # Try to use a WebHook for the "added to queue" message
+      @webHooks.getForChannel(msg.channel, true)
+      .then (hook)=>
+        hook.execSlack @hud.addItemWebhook(msg.guild, qI.requestedBy, qI, queue.items.length+1)
+        msg.delete()
+      # The old method
+      .catch (e)=>
+        msg.channel.sendMessage @hud.addItem msg.guild, qI.requestedBy, qI, queue.items.length+1
+        .then (m)=>
+          omsg = m;
+          setTimeout ->
+            if omsg
+              omsg.delete()
+              omsg = null
+          , 15000
+          msg.delete()
     queue.addToQueue qI
 
 module.exports = PlayerModule
